requirements.txt
python-telegram-bot
static-ffmpeg
